Refactor serializers to their own files
