+++ watch/core/serializers/__init__.py
@@ -1,0 +1,5 @@
+from .item import ItemSerializer
+from .region import RegionSerializer
+from .stacfile import STACFileSerializer
+
+__all__ = ['ItemSerializer', 'RegionSerializer', 'STACFileSerializer']
