+++ watch/core/serializers/item.py
@@ -1,0 +1,5 @@
+from rgd_imagery.stac.serializers import ItemSerializer as ReadOnlyItemSerializer
+
+
+class ItemSerializer(ReadOnlyItemSerializer):
+    pass
