+++ watch/core/serializers/region.py
@@ -1,19 +1,10 @@
 from django.contrib.gis.geos import GeometryCollection, GEOSGeometry
 from django.db import transaction
 from rest_framework import serializers
-from rgd.models import ChecksumFile
-from rgd.serializers import (
-    MODIFIABLE_READ_ONLY_FIELDS,
-    TASK_EVENT_READ_ONLY_FIELDS,
-    ChecksumFileSerializer,
-    RelatedField,
-)
-from rgd_imagery.models import Raster
-from rgd_imagery.serializers import RasterSerializer
 from shapely.geometry import shape
 from shapely.wkb import dumps
 
-from . import models
+from .. import models
 
 
 class RegionSerializer(serializers.BaseSerializer):
@@ -64,24 +55,3 @@ class RegionSerializer(serializers.BaseSerializer):
             feature.end_date = properties['end_date']
             feature.save()
         return instance
-
-
-class STACFileSerializer(serializers.ModelSerializer):
-    # TODO: can this writable?
-    checksumfile = RelatedField(
-        queryset=ChecksumFile.objects.all(), serializer=ChecksumFileSerializer, required=True
-    )
-    # TODO: make sure this is not a required field
-    RelatedField(queryset=Raster.objects.all(), serializer=RasterSerializer, required=False)
-
-    class Meta:
-        model = models.STACFile
-        fields = '__all__'
-        read_only_fields = (
-            MODIFIABLE_READ_ONLY_FIELDS
-            + TASK_EVENT_READ_ONLY_FIELDS
-            + [
-                'processed',
-                'raster',
-            ]
-        )
